requirements.txt
flask
flask-cors
rdflib
python-dotenv